🚧 Improve Focus Mode feature
This commit makes some changes to the behaviour of Focus Mode, so that
it works better and more expectedly; the following points cover the
changes made:

- Added screen dimming - the screen will now heavily dim while the phone
is faced down.
- Added phone usage awareness - the app will detect when the phone is
faced up; this stops the timer indefinitely temporarily.

+++ Liste/Screens/Focus/FMInitialisationViewController.swift
@@ -17,7 +17,7 @@ class FMInitialisationViewController: UIViewController, MSCircularSliderDelegate
     @IBOutlet weak var timeLabel: UILabel!
 
     // MARK: Properties
-    var focusTime: Int = 0
+    var focusTime: Double = 0.0
 
     // MARK: Overrides
     override func viewDidLoad() {
@@ -56,7 +56,7 @@ class FMInitialisationViewController: UIViewController, MSCircularSliderDelegate
         }
 
         self.timeLabel.text = string
-        self.focusTime = time
+        self.focusTime = Double(time)
     }
 
     // MARK: Actions
